Implement GameClock class to replace TimeManager
Create a GameClock class that can be instantiated and manages the passage of time and tick state. This replaces the static TimeManager class that was used previously. The only known difference to the end user right now is that it does not automatically print out tick data to keep the classes decoupled. The facility to do this well can be added, however.